Adds `matchUnion` and `matchTaggedUnion` functions
I added two new functions to match specific types or tags in a `Union`
or `TaggedUnion`, respectively. The functions return `Just` the matched
type if the unions contain the type or tag, and `Nothing` otherwise.

These functions are a lightweight alternative, both syntactically and
performance-wise, to `dissect` when you only need to match on one
branch.