Adds `Ord` instances to `Union` and `TaggedUnion`
We added the instances following the same pattern as the existing `Eq`
instance. This involved adding a function called
`branchDefaultWithIndex` that allows the default result to depend on the
integer index of the branch so that we can correctly return `LT` or `GT`
depending on branch indexes of the unions being compared.

Co-authored-by: Trevis <trevis@flipstone.com>
Co-authored-by: Owen <owen@flipstone.com>